refactor(frontend): flatten message store
