+++ frontend/src/lib/state/room/messages.svelte.ts
@@ -1,3 +1,2 @@
-export { MessageListStore } from './messages/MessageListStore.svelte';
 export { MessagesStore } from './messages/MessagesStore.svelte';
 export { isRootRoomEvent, isThreadEvent } from './messages/filters';
